Move workerpool to utils
